Require, rollup, UMD bundle for JSX/Hyperx and jest. #48.
