+++ hx.js
@@ -1,0 +1,7 @@
+var h = require("./h")
+
+module.exports = {
+    html: require("hyperx")(h),
+    app: require("./app"),
+    h: h
+}
