+++ index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    app: require("./app"),
+    h: require("./h")
+}
