+++ rollup.config.js
@@ -1,11 +1,15 @@
 import buble from "rollup-plugin-buble"
+import uglify from "rollup-plugin-uglify"
 import cjs from "rollup-plugin-commonjs"
 import resolve from "rollup-plugin-node-resolve"
-import uglify from "rollup-plugin-uglify"
 import zopfli from "rollup-plugin-zopfli"
 
 export default {
     plugins: [
-        buble(), cjs(), resolve(), uglify(), zopfli()
+        buble(),
+        cjs(),
+        resolve(),
+        uglify(),
+        zopfli()
     ]
 }
