--- src/html.js
@@ -1,5 +1,0 @@
-import hyperx from "hyperx"
-import app from "./app.js"
-import h from "./h.js"
-
-export default { app, h, html: hyperx(h) }
