--- src/index.js
@@ -1,4 +1,0 @@
-import app from "./app.js"
-import h from "./h.js"
-
-export default { app, h }
